refactor: move RB_Generator and random_circuits in rb utils
